+++ NEWS.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- **README.md**: Use vanilla (not GitHub-flavored) Markdown syntax for snippets
+
 ### Removed
 
 ## [1.4] - 2023-07-11