+++ NEWS.md
@@ -5,7 +5,7 @@ will be documented in this file, in reverse chronological order.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.5] - 2023-07-12
 
 ### Added
 
@@ -86,7 +86,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Initial release, distributed on MELPA.
 
 <!-- bottom -->
-[Unreleased]: https://github.com/ProofGeneral/opam-switch-mode/compare/1.4...HEAD
+[1.5]: https://github.com/ProofGeneral/opam-switch-mode/compare/1.4...1.5
 [1.4]: https://github.com/ProofGeneral/opam-switch-mode/compare/1.3...1.4
 [1.3]: https://github.com/ProofGeneral/opam-switch-mode/compare/1.2...1.3
 [1.2]: https://github.com/ProofGeneral/opam-switch-mode/compare/1.1...1.2